+++ src/main/java/com/jcraft/jsch/jzlib/Tree.java
@@ -205,7 +205,7 @@ final class Tree {
         if (m > max_code)
           continue;
         if (tree[m * 2 + 1] != bits) {
-          s.opt_len += ((long) bits - (long) tree[m * 2 + 1]) * (long) tree[m * 2];
+          s.opt_len += (int) (((long) bits - (long) tree[m * 2 + 1]) * (long) tree[m * 2]);
           tree[m * 2 + 1] = (short) bits;
         }
         n--;
